Revert "Convert to Ts"
This reverts commit 38b1c058565a7ce384631770339cd9ba3489946e.
